♻️ introduce HasuraClient struct
