+++ api/src/lib.rs
@@ -12,7 +12,7 @@ use ::infrastructure::{
 };
 use anyhow::Result;
 use dotenv::dotenv;
-use infrastructure::graphql::hasura_client;
+use infrastructure::graphql::HasuraClient;
 use log::info;
 use rocket::routes;
 use rocket_okapi::swagger_ui::make_swagger_ui;
@@ -37,7 +37,7 @@ pub async fn main() -> Result<()> {
 		event_bus.clone(),
 		AggregateRootRepository::new(database.clone()),
 		AggregateRootRepository::new(database.clone()),
-		Arc::new(hasura_client::new()),
+		Arc::new(HasuraClient::default()),
 	);
 
 	let rocket_handler = rocket::build()
